Move catalog content to its own folder
