+++ frontend/public/components/catalog/_catalog.scss
