+++ frontend/public/components/catalog/catalog-item-icon.tsx
@@ -5,89 +5,89 @@ import * as _ from 'lodash-es';
 import * as classNames from 'classnames';
 
 // eslint-disable-next-line no-unused-vars
-import { K8sResourceKind } from '../module/k8s';
-import * as threeScaleImg from '../imgs/logos/3scale.svg';
-import * as aerogearImg from '../imgs/logos/aerogear.svg';
-import * as amqImg from '../imgs/logos/amq.svg';
-import * as angularjsImg from '../imgs/logos/angularjs.svg';
-import * as ansibleImg from '../imgs/logos/ansible.svg';
-import * as apacheImg from '../imgs/logos/apache.svg';
-import * as beakerImg from '../imgs/logos/beaker.svg';
-import * as capedwarfImg from '../imgs/logos/capedwarf.svg';
-import * as cassandraImg from '../imgs/logos/cassandra.svg';
-import * as clojureImg from '../imgs/logos/clojure.svg';
-import * as codeigniterImg from '../imgs/logos/codeigniter.svg';
-import * as cordovaImg from '../imgs/logos/cordova.png';
-import * as datagridImg from '../imgs/logos/datagrid.svg';
-import * as datavirtImg from '../imgs/logos/datavirt.svg';
-import * as debianImg from '../imgs/logos/debian.svg';
-import * as decisionserverImg from '../imgs/logos/decisionserver.svg';
-import * as djangoImg from '../imgs/logos/django.svg';
-import * as dotnetImg from '../imgs/logos/dotnet.svg';
-import * as drupalImg from '../imgs/logos/drupal.svg';
-import * as eapImg from '../imgs/logos/eap.svg';
-import * as elasticImg from '../imgs/logos/elastic.svg';
-import * as erlangImg from '../imgs/logos/erlang.svg';
-import * as fedoraImg from '../imgs/logos/fedora.svg';
-import * as freebsdImg from '../imgs/logos/freebsd.svg';
-import * as gitImg from '../imgs/logos/git.svg';
-import * as githubImg from '../imgs/logos/github.svg';
-import * as gitlabImg from '../imgs/logos/gitlab.svg';
-import * as glassfishImg from '../imgs/logos/glassfish.svg';
-import * as goGopherImg from '../imgs/logos/go-gopher.svg';
-import * as grailsImg from '../imgs/logos/grails.svg';
-import * as hadoopImg from '../imgs/logos/hadoop.svg';
-import * as haproxyImg from '../imgs/logos/haproxy.svg';
-import * as infinispanImg from '../imgs/logos/infinispan.svg';
-import * as jbossImg from '../imgs/logos/jboss.svg';
-import * as jenkinsImg from '../imgs/logos/jenkins.svg';
-import * as jettyImg from '../imgs/logos/jetty.svg';
-import * as joomlaImg from '../imgs/logos/joomla.svg';
-import * as jrubyImg from '../imgs/logos/jruby.svg';
-import * as jsImg from '../imgs/logos/js.svg';
-import * as kubevirtImg from '../imgs/logos/kubevirt.svg';
-import * as laravelImg from '../imgs/logos/laravel.svg';
-import * as loadBalancerImg from '../imgs/logos/load-balancer.svg';
-import * as mariadbImg from '../imgs/logos/mariadb.svg';
-import * as mediawikiImg from '../imgs/logos/mediawiki.svg';
-import * as memcachedImg from '../imgs/logos/memcached.svg';
-import * as mongodbImg from '../imgs/logos/mongodb.svg';
-import * as mssqlImg from '../imgs/logos/mssql.svg';
-import * as mysqlDatabaseImg from '../imgs/logos/mysql-database.svg';
-import * as nginxImg from '../imgs/logos/nginx.svg';
-import * as nodejsImg from '../imgs/logos/nodejs.svg';
-import * as openjdkImg from '../imgs/logos/openjdk.svg';
-import * as openshiftImg from '../imgs/logos/openshift.svg';
-import * as openstackImg from '../imgs/logos/openstack.svg';
-import * as otherLinuxImg from '../imgs/logos/other-linux.svg';
-import * as otherUnknownImg from '../imgs/logos/other-unknown.svg';
-import * as perlImg from '../imgs/logos/perl.svg';
-import * as phalconImg from '../imgs/logos/phalcon.svg';
-import * as phpImg from '../imgs/logos/php.svg';
-import * as playImg from '../imgs/logos/play.svg';
-import * as postgresqlImg from '../imgs/logos/postgresql.svg';
-import * as processserverImg from '../imgs/logos/processserver.svg';
-import * as pythonImg from '../imgs/logos/python.svg';
-import * as rabbitmqImg from '../imgs/logos/rabbitmq.svg';
-import * as railsImg from '../imgs/logos/rails.svg';
-import * as redisImg from '../imgs/logos/redis.svg';
-import * as rhIntegrationImg from '../imgs/logos/rh-integration.svg';
-import * as rhTomcatImg from '../imgs/logos/rh-tomcat.svg';
-import * as rubyImg from '../imgs/logos/ruby.svg';
-import * as scalaImg from '../imgs/logos/scala.svg';
-import * as shadowmanImg from '../imgs/logos/shadowman.svg';
-import * as springImg from '../imgs/logos/spring.svg';
-import * as ssoImg from '../imgs/logos/sso.svg';
-import * as stackoverflowImg from '../imgs/logos/stackoverflow.svg';
-import * as suseImg from '../imgs/logos/suse.svg';
-import * as symfonyImg from '../imgs/logos/symfony.svg';
-import * as tomcatImg from '../imgs/logos/tomcat.svg';
-import * as ubuntuImg from '../imgs/logos/ubuntu.svg';
-import * as wildflyImg from '../imgs/logos/wildfly.svg';
-import * as windowsImg from '../imgs/logos/windows.svg';
-import * as wordpressImg from '../imgs/logos/wordpress.svg';
-import * as xamarinImg from '../imgs/logos/xamarin.svg';
-import * as zendImg from '../imgs/logos/zend.svg';
+import { K8sResourceKind } from '../../module/k8s';
+import * as threeScaleImg from '../../imgs/logos/3scale.svg';
+import * as aerogearImg from '../../imgs/logos/aerogear.svg';
+import * as amqImg from '../../imgs/logos/amq.svg';
+import * as angularjsImg from '../../imgs/logos/angularjs.svg';
+import * as ansibleImg from '../../imgs/logos/ansible.svg';
+import * as apacheImg from '../../imgs/logos/apache.svg';
+import * as beakerImg from '../../imgs/logos/beaker.svg';
+import * as capedwarfImg from '../../imgs/logos/capedwarf.svg';
+import * as cassandraImg from '../../imgs/logos/cassandra.svg';
+import * as clojureImg from '../../imgs/logos/clojure.svg';
+import * as codeigniterImg from '../../imgs/logos/codeigniter.svg';
+import * as cordovaImg from '../../imgs/logos/cordova.png';
+import * as datagridImg from '../../imgs/logos/datagrid.svg';
+import * as datavirtImg from '../../imgs/logos/datavirt.svg';
+import * as debianImg from '../../imgs/logos/debian.svg';
+import * as decisionserverImg from '../../imgs/logos/decisionserver.svg';
+import * as djangoImg from '../../imgs/logos/django.svg';
+import * as dotnetImg from '../../imgs/logos/dotnet.svg';
+import * as drupalImg from '../../imgs/logos/drupal.svg';
+import * as eapImg from '../../imgs/logos/eap.svg';
+import * as elasticImg from '../../imgs/logos/elastic.svg';
+import * as erlangImg from '../../imgs/logos/erlang.svg';
+import * as fedoraImg from '../../imgs/logos/fedora.svg';
+import * as freebsdImg from '../../imgs/logos/freebsd.svg';
+import * as gitImg from '../../imgs/logos/git.svg';
+import * as githubImg from '../../imgs/logos/github.svg';
+import * as gitlabImg from '../../imgs/logos/gitlab.svg';
+import * as glassfishImg from '../../imgs/logos/glassfish.svg';
+import * as goGopherImg from '../../imgs/logos/go-gopher.svg';
+import * as grailsImg from '../../imgs/logos/grails.svg';
+import * as hadoopImg from '../../imgs/logos/hadoop.svg';
+import * as haproxyImg from '../../imgs/logos/haproxy.svg';
+import * as infinispanImg from '../../imgs/logos/infinispan.svg';
+import * as jbossImg from '../../imgs/logos/jboss.svg';
+import * as jenkinsImg from '../../imgs/logos/jenkins.svg';
+import * as jettyImg from '../../imgs/logos/jetty.svg';
+import * as joomlaImg from '../../imgs/logos/joomla.svg';
+import * as jrubyImg from '../../imgs/logos/jruby.svg';
+import * as jsImg from '../../imgs/logos/js.svg';
+import * as kubevirtImg from '../../imgs/logos/kubevirt.svg';
+import * as laravelImg from '../../imgs/logos/laravel.svg';
+import * as loadBalancerImg from '../../imgs/logos/load-balancer.svg';
+import * as mariadbImg from '../../imgs/logos/mariadb.svg';
+import * as mediawikiImg from '../../imgs/logos/mediawiki.svg';
+import * as memcachedImg from '../../imgs/logos/memcached.svg';
+import * as mongodbImg from '../../imgs/logos/mongodb.svg';
+import * as mssqlImg from '../../imgs/logos/mssql.svg';
+import * as mysqlDatabaseImg from '../../imgs/logos/mysql-database.svg';
+import * as nginxImg from '../../imgs/logos/nginx.svg';
+import * as nodejsImg from '../../imgs/logos/nodejs.svg';
+import * as openjdkImg from '../../imgs/logos/openjdk.svg';
+import * as openshiftImg from '../../imgs/logos/openshift.svg';
+import * as openstackImg from '../../imgs/logos/openstack.svg';
+import * as otherLinuxImg from '../../imgs/logos/other-linux.svg';
+import * as otherUnknownImg from '../../imgs/logos/other-unknown.svg';
+import * as perlImg from '../../imgs/logos/perl.svg';
+import * as phalconImg from '../../imgs/logos/phalcon.svg';
+import * as phpImg from '../../imgs/logos/php.svg';
+import * as playImg from '../../imgs/logos/play.svg';
+import * as postgresqlImg from '../../imgs/logos/postgresql.svg';
+import * as processserverImg from '../../imgs/logos/processserver.svg';
+import * as pythonImg from '../../imgs/logos/python.svg';
+import * as rabbitmqImg from '../../imgs/logos/rabbitmq.svg';
+import * as railsImg from '../../imgs/logos/rails.svg';
+import * as redisImg from '../../imgs/logos/redis.svg';
+import * as rhIntegrationImg from '../../imgs/logos/rh-integration.svg';
+import * as rhTomcatImg from '../../imgs/logos/rh-tomcat.svg';
+import * as rubyImg from '../../imgs/logos/ruby.svg';
+import * as scalaImg from '../../imgs/logos/scala.svg';
+import * as shadowmanImg from '../../imgs/logos/shadowman.svg';
+import * as springImg from '../../imgs/logos/spring.svg';
+import * as ssoImg from '../../imgs/logos/sso.svg';
+import * as stackoverflowImg from '../../imgs/logos/stackoverflow.svg';
+import * as suseImg from '../../imgs/logos/suse.svg';
+import * as symfonyImg from '../../imgs/logos/symfony.svg';
+import * as tomcatImg from '../../imgs/logos/tomcat.svg';
+import * as ubuntuImg from '../../imgs/logos/ubuntu.svg';
+import * as wildflyImg from '../../imgs/logos/wildfly.svg';
+import * as windowsImg from '../../imgs/logos/windows.svg';
+import * as wordpressImg from '../../imgs/logos/wordpress.svg';
+import * as xamarinImg from '../../imgs/logos/xamarin.svg';
+import * as zendImg from '../../imgs/logos/zend.svg';
 
 const logos = new Map()
   .set('icon-3scale', threeScaleImg)
