+++ frontend/public/components/cluster-service-class-info.tsx
@@ -5,7 +5,7 @@ import * as _ from 'lodash-es';
 
 // eslint-disable-next-line no-unused-vars
 import { K8sResourceKind, serviceClassDisplayName } from '../module/k8s';
-import { ClusterServiceClassIcon } from './catalog-item-icon';
+import { ClusterServiceClassIcon } from './catalog/catalog-item-icon';
 
 export const ClusterServiceClassInfo: React.SFC<ClusterServiceClassInfoProps> = ({obj: serviceClass}) => {
   const displayName = serviceClassDisplayName(serviceClass);
