+++ frontend/public/components/cluster-service-class.tsx
@@ -10,7 +10,7 @@ import { viewYamlComponent } from './utils/horizontal-nav';
 import { ClusterServiceClassModel } from '../models';
 // eslint-disable-next-line no-unused-vars
 import { K8sResourceKind, K8sResourceKindReference, serviceClassDisplayName } from '../module/k8s';
-import { ClusterServiceClassIcon } from './catalog-item-icon';
+import { ClusterServiceClassIcon } from './catalog/catalog-item-icon';
 import { ClusterServicePlanPage } from './cluster-service-plan';
 import { ClusterServiceClassInfo } from './cluster-service-class-info';
